+++ setup/runvenv.bat
@@ -1,0 +1,2 @@
+@echo off
+cmd /k "cd /d C:\Users\johnil1\Desktop\Studio_5_FDAS\venv\Scripts & activate"
